refactor(version): extract nodejs resolution to nodejs.go
Move Node.js LTS version resolution to a dedicated file.
